Interpolation milestone
It's a bit janky, but parsing interpolation finally works! The lexer now has a _mode stack_, which is used by the parser to change the current lexer mode. This allows the parser to change into a special mode when it need to parse a double-quoted string.

The janky bit is the peek/advance logic in the parser itself (the lexer is quite elegant), so that aspect of the parser needs a lot of cleanup.