Change block named parameter syntax
The previous syntax used square brackets, but square brackets are more commonly associated with list and map literals in other languages. Because of this, and because Riptide lacks (and currently needs) its own map and list literal syntax, square brackets will now be reserved for that.

To put minimal stylistic change on block syntax, angle brackets are used instead. This does mean that angle brackets can now no longer be allowed in identifiers or symbols.