Add list and table literals
Since square brackets are now no longer reserved, use them to implement rudimentary list and table literals. Both these syntaxes use square brackets, with tables relying on a short lookahead to look for the `:` character. Lists look like this:

    [a b c]

while maps look like this:

    [a:b c:d]

I'm not sure if commas would help or hurt the syntax or not, considering that commas are not used anywhere else.

This implementation has an issue with newlines. For now newlines must be escaped:

    [ \
        a: b \
        c: d \
    ]